+++ src/include/rg_rand.h
@@ -1,2 +1,4 @@
-long longrand(long randomnum);
-long slongrand(unsigned long seed);
+#include <stdint.h>
+
+int32_t longrand(int32_t randomnum);
+int32_t slongrand(uint32_t seed);
